[service] Move on_target to a mixin
